requirements.txt
Flask
flask-cors
pyinstaller
praat-parselmouth
numpy
matplotlib
seaborn
tornado